Updating auction.go

+++ adserver/auction.go
@@ -12,7 +12,7 @@ import (
 )
 
 var (
-	NewAdImpressionThreshold = flag.Int("newAdTreshold", 5, "Impression threshold for considering an ad as new")
+	NewAdImpressionThreshold = flag.Int64("newAdTreshold", 5, "Impression threshold for considering an ad as new")
 	NewAdSelectionProbability = flag.Float64("newAdProb", 0.25, "Probability of selecting a new ad")
 	ExperiencedAdSelectionProbability = flag.Float64("expAdProb", 0.75, "Probability of selecting a exprienced ad")
 )
